refactor: render the upsell screens from the existing Pro previews
The upsell pages were drawing their own mock in PHP, which meant two
previews per feature that could drift apart. Each screen now mounts the
ProPreviews panel the settings page already shows, so there is one
preview per feature and one place to change it.

That removes the hand-written markup, the bespoke styles and the base
class they needed; what stays is the menu entry, its Pro badge, the
position Pro will later take, and a mount point.
